requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==41.0.1
decouple==0.0.7
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-countries==7.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
humanfriendly==10.0
humanize==4.6.0
idna==3.4
Markdown==3.4.3
marshmallow==3.17.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
pyreadline3==3.4.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rut-chile==2.0.1
sqlparse==0.4.4
stripe==5.4.0
transbank-sdk==4.0.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
